+++ R/mda.R
@@ -469,3 +469,19 @@ findlipid <-
                 list$RKMD <- cbind.data.frame(t,c)
                 return(list)
         }
+
+#' Find PFCs based on mass defect analysis
+#' @param list list with data as peaks list, mz, rt and group information, retention time should be in seconds
+#' @param mode 'pos' for positive mode, 'neg' for negative mode and 'none' for neutral mass, only support [M+H] and [M-H] for each mode
+#' @return list list with potential PFCs compounds index
+#' @references Liu, Y.; D’Agostino, L. A.; Qu, G.; Jiang, G.; Martin, J. W. High-Resolution Mass Spectrometry (HRMS) Methods for Nontarget Discovery and Characterization of Poly- and per-Fluoroalkyl Substances (PFASs) in Environmental and Human Samples. TrAC Trends in Analytical Chemistry 2019, 121, 115420.
+#' @examples
+#' data(list)
+#' pfc <- findpfc(list)
+#' @export
+findpfc <- function(list){
+        md <- list$mz-round(list$mz)
+        # mass defect -0.1 to 0.15
+        list$pfc <- md > -0.1 & md < 0.15
+        return(list)
+}